+++ Packages/com.unity.inputsystem/InputSystem/Plugins/PlayerInput/PlayerInput.cs
@@ -402,14 +402,13 @@ namespace UnityEngine.InputSystem.Plugins.PlayerInput
             set => m_Camera = value;
         }
 
-        //nuke?
         /// <summary>
-        /// The event system that should be fed with UI events from the player's inputs.
+        /// UI InputModule that should have it's input actions synchronized to this PlayerInput's actions.
         /// </summary>
-        public EventSystem uiEventSystem
+        public InputSystemUIInputModule uiInputModule
         {
-            get { return m_UIEventSystem; }
-            set { throw new NotImplementedException(); }
+            get { return m_UIInputModule; }
+            set { m_UIInputModule = value; }
         }
 
         /// <summary>
@@ -459,24 +458,24 @@ namespace UnityEngine.InputSystem.Plugins.PlayerInput
             {
                 var actionMap = m_Actions.TryGetActionMap(m_DefaultActionMap);
                 if (actionMap != null)
+                {
                     actionMap.Enable();
+                    m_EnabledActionMap = actionMap;
+                }
                 else
                     Debug.LogError($"Cannot find action map '{m_DefaultActionMap}' in '{m_Actions}'", this);
             }
-
             m_InputActive = true;
         }
 
         public void PassivateInput()
         {
-            // Disable all enabled action maps.
-            if (m_Actions != null)
-                m_Actions.Disable();
+            m_EnabledActionMap?.Disable();
 
             m_InputActive = false;
         }
 
-        public void SwitchActions(string mapNameOrId)
+        public void SwitchCurrentActionMap(string mapNameOrId)
         {
             // Must be enabled.
             if (!m_Enabled)
@@ -500,8 +499,7 @@ namespace UnityEngine.InputSystem.Plugins.PlayerInput
                 return;
             }
 
-            m_Actions.Disable();
-            actionMap.Enable();
+            currentActionMap = actionMap;
         }
 
         public static PlayerInput GetPlayerByIndex(int playerIndex)
@@ -611,8 +609,8 @@ namespace UnityEngine.InputSystem.Plugins.PlayerInput
         [SerializeField] internal InputActionAsset m_Actions;
         [Tooltip("Determine how notifications should be sent when an input-related event associated with the player happens.")]
         [SerializeField] internal PlayerNotifications m_NotificationBehavior;
-        [Tooltip("UI EventSystem that should receive input from the actions associated with the player. TODO")]
-        [SerializeField] internal EventSystem m_UIEventSystem;
+        [Tooltip("UI InputModule that should have it's input actions synchronized to this PlayerInput's actions.")]
+        [SerializeField] internal InputSystemUIInputModule m_UIInputModule;
         [Tooltip("Event that is triggered when the PlayerInput loses a paired device (e.g. its battery runs out).")]
         [SerializeField] internal DeviceLostEvent m_DeviceLostEvent;
         [SerializeField] internal DeviceRegainedEvent m_DeviceRegainedEvent;
@@ -628,6 +626,19 @@ namespace UnityEngine.InputSystem.Plugins.PlayerInput
         // Value object we use when sending messages via SendMessage() or BroadcastMessage(). Can be ignored
         // by the receiver. We reuse the same object over and over to avoid allocating garbage.
         [NonSerialized] private InputValue m_InputValueObject;
+
+        [NonSerialized] internal InputActionMap m_EnabledActionMap;
+
+        public InputActionMap currentActionMap
+        {
+            get => m_EnabledActionMap;
+            set
+            {
+                m_EnabledActionMap?.Disable();
+                m_EnabledActionMap = value;
+                m_EnabledActionMap?.Enable();
+            }
+        }
 
         [NonSerialized] private int m_PlayerIndex = -1;
         [NonSerialized] private bool m_InputActive;
@@ -667,6 +678,9 @@ namespace UnityEngine.InputSystem.Plugins.PlayerInput
                     m_Actions = Instantiate(m_Actions);
                     break;
                 }
+
+            if (uiInputModule != null)
+                uiInputModule.actionsAsset = m_Actions;
 
             switch (m_NotificationBehavior)
             {