+++ Packages/com.unity.inputsystem/InputSystem/Plugins/XR/Vive.cs
@@ -97,11 +97,10 @@ namespace UnityEngine.InputSystem.Plugins.XR
     [InputControlLayout]
     public class KnucklesController : XRControllerWithRumble
     {
- 
         [InputControl(aliases = new[] { "B",  "Primary"})]
         public ButtonControl primaryButton { get; private set; }
 
-        [InputControl(aliases = new[] { "JoystickOrPadPressed" })]        
+        [InputControl(aliases = new[] { "JoystickOrPadPressed" })]
         public ButtonControl trackpadPressed { get; private set; }
         [InputControl(aliases = new[] { "JoystickOrPadTouched" })]
         public ButtonControl trackpadTouched { get; private set; }
@@ -125,14 +124,12 @@ namespace UnityEngine.InputSystem.Plugins.XR
         public Vector3Control deviceAngularVelocity { get; private set; }
 
 
-
-
         protected override void FinishSetup(InputDeviceBuilder builder)
         {
             base.FinishSetup(builder);
-                    
+
             gripPressed = builder.GetControl<ButtonControl>("gripPressed");
-            primaryButton = builder.GetControl<ButtonControl>("primary");            
+            primaryButton = builder.GetControl<ButtonControl>("primary");
             trackpadPressed = builder.GetControl<ButtonControl>("trackpadPressed");
             trackpadTouched = builder.GetControl<ButtonControl>("trackpadTouched");
             trackpad = builder.GetControl<Vector2Control>("trackpad");